chore: Split up publish and validate
