+++ api.planx.uk/modules/flows/validate/service.ts
@@ -1,16 +1,13 @@
 import * as jsondiffpatch from "jsondiffpatch";
 import { Request, Response, NextFunction } from "express";
 import { dataMerged, getMostRecentPublishedFlow } from "../../../helpers";
-import { gql } from "graphql-request";
 import intersection from "lodash/intersection";
 import {
   ComponentType,
   FlowGraph,
   Node,
 } from "@opensystemslab/planx-core/types";
-import { userContext } from "../../auth/middleware";
 import type { Entry } from "type-fest";
-import { getClient } from "../../../client";
 
 const validateAndDiffFlow = async (
   req: Request,
@@ -53,85 +50,6 @@ const validateAndDiffFlow = async (
       const alteredNodes = Object.keys(delta).map((key) => ({
         id: key,
         ...flattenedFlow[key],
-      }));
-
-      return res.json({
-        alteredNodes,
-      });
-    } else {
-      return res.json({
-        alteredNodes: null,
-        message: "No new changes to publish",
-      });
-    }
-  } catch (error) {
-    return next(error);
-  }
-};
-
-interface PublishFlow {
-  publishedFlow: {
-    id: string;
-    flowId: string;
-    publisherId: string;
-    createdAt: string;
-    data: FlowGraph;
-  };
-}
-
-const publishFlow = async (
-  req: Request,
-  res: Response,
-  next: NextFunction,
-): Promise<Response | NextFunction | void> => {
-  try {
-    const flattenedFlow = await dataMerged(req.params.flowId);
-    const mostRecent = await getMostRecentPublishedFlow(req.params.flowId);
-    const delta = jsondiffpatch.diff(mostRecent, flattenedFlow);
-
-    const userId = userContext.getStore()?.user?.sub;
-    if (!userId) throw Error("User details missing from request");
-
-    if (delta) {
-      const { client: $client } = getClient();
-      const response = await $client.request<PublishFlow>(
-        gql`
-          mutation PublishFlow(
-            $data: jsonb = {}
-            $flow_id: uuid
-            $publisher_id: Int
-            $summary: String
-          ) {
-            publishedFlow: insert_published_flows_one(
-              object: {
-                data: $data
-                flow_id: $flow_id
-                publisher_id: $publisher_id
-                summary: $summary
-              }
-            ) {
-              id
-              flowId: flow_id
-              publisherId: publisher_id
-              createdAt: created_at
-              data
-            }
-          }
-        `,
-        {
-          data: flattenedFlow,
-          flow_id: req.params.flowId,
-          publisher_id: parseInt(userId),
-          summary: req.query?.summary || null,
-        },
-      );
-
-      const publishedFlow =
-        response.publishedFlow && response.publishedFlow.data;
-
-      const alteredNodes = Object.keys(delta).map((key) => ({
-        id: key,
-        ...publishedFlow[key],
       }));
 
       return res.json({
@@ -320,4 +238,4 @@ const numberOfComponentType = (
   return nodeIds?.length;
 };
 
-export { validateAndDiffFlow, publishFlow };
+export { validateAndDiffFlow };
